fix: 로그아웃 / 회원탈퇴 후 재로그인 시 닉네임 잘못 받아오는 오류 수정
